review

+++ wemake-server/internal/service/order/order_review_service.go
@@ -102,7 +102,7 @@ func (s *OrderService) CreateReview(orderID, userID int64, role string, input Cr
 		FactoryID: order.FactoryID,
 		UserID:    userID,
 		OrderID:   &orderIDPtr,
-		Rating:    input.Rating,
+		Rating:    float64(input.Rating),
 		Comment:   comment,
 		ImageURLs: imageURLs,
 	}
@@ -127,7 +127,7 @@ func (s *OrderService) CreateReview(orderID, userID int64, role string, input Cr
 		UserID:  order.FactoryID,
 		Type:    "REVIEW_RECEIVED",
 		Title:   "ได้รับรีวิวใหม่",
-		Message: fmt.Sprintf("ลูกค้าให้ %d ดาว: \"%s\"", review.Rating, helper.TrimNotificationPreview(review.Comment, 80)),
+		Message: fmt.Sprintf("ลูกค้าให้ %d ดาว: \"%s\"", int(review.Rating), helper.TrimNotificationPreview(review.Comment, 80)),
 		LinkTo:  helper.FactoryOrderLink(orderID),
 		Data: helper.NotificationData(map[string]interface{}{
 			"review_id": review.ReviewID,
